More one time fixed chat issues
Fixed date bug: It used to desapear when ListView detected a need of rendering

Fixed time issue with zeros before single numbers. It was showing only '1:5' instead '01:05'

new way to send data. now it is sended as a message to users. This is nice because can be done by the server